FIX stop autoload failure on case sensitive system
Trying to install and run cow on an Ubuntu system I was greeted by a fatal error "can't find class ChangeLog" - which was caused by the differences between the class name call of `ChangeLog` and the class name (and thus filename) being `Changelog.php`. Correcting this single erroneous reference has fixed the issue.