API Remove archive upload functionality
See https://github.com/silverstripe/silverstripe-framework/issues/9232.
The "wait" step was somewhat coupled to the archive functionality.
While it could provide some value to hold off the command success
until the new versions are available in Packagist, there's no longer
a need for it: We're not trying to *use* the new versions immediately
through a "composer install" for the archive build.

In general, I think we can rely on Packagist loading in new versions
within a few minutes, and this happens consistently enough
that we don't need a specific "wait" step any more.